Extract files_to_prompt from Gemini action
