+++ tests/components/google_generative_ai_conversation/snapshots/test_init.ambr
@@ -122,8 +122,8 @@
       dict({
         'contents': list([
           'Describe this image from my doorbell camera',
-          b'some file',
-          b'some file',
+          File(name='doorbell_snapshot.jpg', display_name=None, mime_type=None, size_bytes=None, create_time=None, expiration_time=None, update_time=None, sha256_hash=None, uri=None, download_uri=None, state=<FileState.ACTIVE: 'ACTIVE'>, source=None, video_metadata=None, error=None),
+          File(name='context.txt', display_name=None, mime_type=None, size_bytes=None, create_time=None, expiration_time=None, update_time=None, sha256_hash=None, uri=None, download_uri=None, state=<FileState.ACTIVE: 'ACTIVE'>, source=None, video_metadata=None, error=None),
         ]),
         'model': 'models/gemini-2.5-flash',
       }),
